requirements.txt
streamlit
llama-index.core
openai
python-docx
PyPDF2